requirements.txt
langchain
langchain-community
supabase
python-dotenv
langchain-openai
transformers 
sentence-transformers
langchain_huggingface
langgraph
langchain-google-genai
